Create data directory and move data classes
