+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
-import Quiz from './Quiz';
+import Quiz from './data/Quiz';
 
 function App() {
   return (
